refactor: move skyblock package to hypixel.skyblock package
