+++ src/main/kotlin/net/llvg/exec/utils/hypixel/skyblock/catacombs/map/scan/CatacombsMap.kt
@@ -17,9 +17,9 @@
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.llvg.exec.utils.skyblock.catacombs.map.scan
+package net.llvg.exec.utils.hypixel.skyblock.catacombs.map.scan
 
-import net.llvg.exec.utils.skyblock.catacombs.map.RoomData
+import net.llvg.exec.utils.hypixel.skyblock.catacombs.map.RoomData
 import net.llvg.exec.utils.vector.Vec2I
 
 @Suppress("MemberVisibilityCanBePrivate")
